Доработал стили
