+++ packages/client/src/theme/palette.ts
@@ -5,14 +5,23 @@ export const getDesignTokens = (mode: PaletteMode): PaletteOptions => ({
   mode,
   ...(mode === 'light'
     ? {
-        primary: {
-          main: '#3968DA',
-        },
-        layout: {
-          fontColor: '#273443',
-          pageBackgroundColor: '#F3F8FE',
-          headerAndFooterBackgroundColor: '#E9F3FF',
-        },
-      }
-    : {}),
+      primary: {
+        main: '#3968DA',
+      },
+      layout: {
+        fontColor: '#273443',
+        pageBackgroundColor: '#F3F8FE',
+        headerAndFooterBackgroundColor: '#E9F3FF'
+      },
+    }
+    : { // TODO: добавить стили для темной темы
+      primary: {
+        main: '#3968DA',
+      },
+      layout: {
+        fontColor: '#273443',
+        pageBackgroundColor: '#F3F8FE',
+        headerAndFooterBackgroundColor: '#E9F3FF'
+      },
+    }),
 })
